requirements.txt
# Install with: 
# conda activate your_env_name
# python -m pip install -r requirements.txt
yfinance>=0.2.0
pandas>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
